Remove ballots index action
For starters, the navigation on this page was broken. Instead of trying
to fix a page that provides no value it is better to just remove this
(which will also allow us to remove our Chartkick dependency) and return
to it later if we believe we can find value in it.