Require users to sign in before voting
Currently if a user is not signed in and selects "Vote!" and error
occurs when the controller tries to access the current user. This will
simply redirect the user to the log in page with a message saying they
must sign in or sign up. After doing so, they will be redirected to the
voting page.